chore(ImageModal): convert to vue 3 and update test and docs
